Added returns to every function
Because I am a very silly sausage, I forgot to put return statements in the functions, making it impossible to tell if a function was successful. This caused them to print a message to BizHawk's screen even if the function failed. I have fixed this and added the needed returns.